requirements.txt
py-cord>=2.6.0
audioop-lts>=0.2.1   # backport of audioop removed in Python 3.13
paho-mqtt>=1.6.1
python-dotenv>=1.0.0
requests>=2.31.0
